refactor(printf): extract shared types into specifier module;
Splits the printf module into single-purpose files:
- specifier: shared type definitions + ResolvedSpecifier
- parse: PrintfParser iterator + PrintfItem
- format: Formatter, Radix, & DigitBuffer
- mod: entry points only
