+++ src/libc/str/printf/format.rs
@@ -3,10 +3,7 @@ use std::{
     io::Write,
 };
 
-use super::{
-    parse::{Conversion, LengthModifier, PadMode},
-    ResolvedSpecifier,
-};
+use super::specifier::{Conversion, LengthModifier, PadMode, ResolvedSpecifier};
 
 pub struct Formatter<W: Write> {
     writer: W,
@@ -365,27 +362,5 @@ impl DigitBuffer {
 
     fn digits(&self) -> &[u8] {
         &self.storage[..self.length]
-    }
-}
-
-impl LengthModifier {
-    pub unsafe fn extract_signed(self, args: &mut VaList<'_, '_>) -> i64 {
-        match self {
-            Self::HalfHalf => (args.arg::<i32>() as i8) as i64,
-            Self::Half => (args.arg::<i32>() as i16) as i64,
-            Self::None | Self::LongDouble => args.arg::<i32>() as i64,
-            Self::Long | Self::LongLong | Self::IntMax => args.arg::<i64>(),
-            Self::Size | Self::Ptrdiff => args.arg::<isize>() as i64,
-        }
-    }
-
-    pub unsafe fn extract_unsigned(self, args: &mut VaList<'_, '_>) -> u64 {
-        match self {
-            Self::HalfHalf => (args.arg::<u32>() as u8) as u64,
-            Self::Half => (args.arg::<u32>() as u16) as u64,
-            Self::None | Self::LongDouble => args.arg::<u32>() as u64,
-            Self::Long | Self::LongLong | Self::IntMax => args.arg::<u64>(),
-            Self::Size | Self::Ptrdiff => args.arg::<usize>() as u64,
-        }
     }
 }
