+++ src/libc/str/printf/parse.rs
@@ -1,156 +1,6 @@
 use std::ffi::CStr;
 
-/// Raw flags parsed from a printf format specifier.
-/// No interpretation is applied — flag validity and interactions
-/// are resolved later by `ResolvedSpecifier`.
-#[derive(Default, Clone, Copy)]
-pub struct RawFlags {
-    pub left_justify: bool,
-    pub force_sign: bool,
-    pub space_sign: bool,
-    pub alternate: bool,
-    pub zero_pad: bool,
-}
-
-#[derive(Clone, Copy)]
-pub enum SignMode {
-    None,
-    Space,
-    Force,
-}
-
-impl SignMode {
-    pub fn sign_byte(self, is_negative: bool) -> Option<u8> {
-        if is_negative {
-            Some(b'-')
-        } else {
-            match self {
-                Self::Force => Some(b'+'),
-                Self::Space => Some(b' '),
-                Self::None => None,
-            }
-        }
-    }
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-pub enum PadMode {
-    RightAlign,
-    LeftAlign,
-    ZeroPad,
-}
-
-impl PadMode {
-    /// Resolve where padding goes: `(left_spaces, extra_zeros, right_spaces)`.
-    ///
-    /// For zero-pad, the padding becomes additional zeros between sign/prefix and digits.
-    /// For right-justify (default), padding is leading spaces.
-    /// For left-justify, padding is trailing spaces.
-    pub fn resolve_padding(self, padding: usize) -> (usize, usize, usize) {
-        match self {
-            Self::ZeroPad => (0, padding, 0),
-            Self::LeftAlign => (0, 0, padding),
-            Self::RightAlign => (padding, 0, 0),
-        }
-    }
-}
-
-/// Length modifier (`h`, `hh`, `l`, `ll`, `z`, `t`, `j`).
-#[derive(Default, Clone, Copy, PartialEq, Eq)]
-pub enum LengthModifier {
-    #[default]
-    None,
-    HalfHalf,
-    Half,
-    Long,
-    LongLong,
-    LongDouble,
-    Size,
-    Ptrdiff,
-    IntMax,
-}
-
-#[derive(Default, Clone, Copy)]
-pub enum DimensionSpecifier {
-    #[default]
-    Unspecified,
-    Fixed(usize),
-    FromNextArg,
-}
-
-#[derive(Clone, Copy)]
-pub enum Conversion {
-    SignedInt,
-    UnsignedInt,
-    Octal,
-    Hex { uppercase: bool },
-    Float(FloatFormat),
-    String,
-    Char,
-    Pointer,
-    CharCount,
-}
-
-impl Conversion {
-    fn from_byte(byte: u8) -> Option<Self> {
-        match byte {
-            b'd' | b'i' => Some(Self::SignedInt),
-            b'u' => Some(Self::UnsignedInt),
-            b'o' => Some(Self::Octal),
-            b'x' => Some(Self::Hex { uppercase: false }),
-            b'X' => Some(Self::Hex { uppercase: true }),
-            b'f' => Some(Self::Float(FloatFormat::Fixed { uppercase: false })),
-            b'F' => Some(Self::Float(FloatFormat::Fixed { uppercase: true })),
-            b'e' => Some(Self::Float(FloatFormat::Scientific { uppercase: false })),
-            b'E' => Some(Self::Float(FloatFormat::Scientific { uppercase: true })),
-            b'g' => Some(Self::Float(FloatFormat::General { uppercase: false })),
-            b'G' => Some(Self::Float(FloatFormat::General { uppercase: true })),
-            b'a' => Some(Self::Float(FloatFormat::Hex { uppercase: false })),
-            b'A' => Some(Self::Float(FloatFormat::Hex { uppercase: true })),
-            b's' => Some(Self::String),
-            b'c' => Some(Self::Char),
-            b'p' => Some(Self::Pointer),
-            b'n' => Some(Self::CharCount),
-            _ => None,
-        }
-    }
-
-    pub fn is_signed(self) -> bool {
-        matches!(self, Self::SignedInt | Self::Float(_))
-    }
-
-    pub fn is_float(self) -> bool {
-        matches!(self, Self::Float(_))
-    }
-
-    pub fn is_integer(self) -> bool {
-        matches!(
-            self,
-            Self::SignedInt | Self::UnsignedInt | Self::Octal | Self::Hex { .. }
-        )
-    }
-
-    pub fn is_numeric(self) -> bool {
-        self.is_integer() || matches!(self, Self::Float(_))
-    }
-}
-
-#[derive(Clone, Copy)]
-pub enum FloatFormat {
-    Fixed { uppercase: bool },
-    Scientific { uppercase: bool },
-    General { uppercase: bool },
-    Hex { uppercase: bool },
-}
-
-/// A parsed format specifier — everything between `%` and the conversion character.
-pub struct PrintfSpecifier {
-    pub flags: RawFlags,
-    pub width: DimensionSpecifier,
-    pub precision: DimensionSpecifier,
-    pub length: LengthModifier,
-    pub conversion: Conversion,
-}
+use super::specifier::{Conversion, DimensionSpecifier, LengthModifier, PrintfSpecifier, RawFlags};
 
 /// An item yielded by [`PrintfParser`].
 pub enum PrintfItem<'a> {
